requirements.txt
streamlit
langchain
langchain-openai
langchain-google-genai
langchain-anthropic
pypdf
faiss-cpu
python-dotenv